Clean up code, change method of convolution and allow 4 channels
Some code clean up, remove comments and debug print satements

Added a sharpen and remove noise kernel. And used this to modify how
the final image is being obtained

Removed the useage of the consant 3, as this was the hard coded image
channels. Now it uses the ones gathered from stb_image, which means
images with 3 or 4 channels will now work